Changed outer catch to log exceptions.
I changed the outermost try/catch block so that if an unhandled exception occurs, the contents are not only written to the standard error device, but they are also written to the log file if the -log option is used.  This helps to ensure that stack traces show up in logs if there's an unexpected exception.